refactor(recent-food): move RemoveFromRecentButton to recent-food module
