+++ src/modules/recent-food/ui/RemoveFromRecentButton.tsx
